extract test run from test spec to support before and after
